requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
pydantic==2.7.1
python-multipart==0.0.9
python-socketio
grpcio

passlib==1.7.4
passlib[bcrypt]
PyJWT[crypto]

requests==2.32.2
aiohttp==3.9.5
httpx

# added libraries
langchain==0.3.25
langchain-core==0.3.63
langchain-community==0.3.24
langchain-text-splitters==0.3.8
langchain-ollama==0.3.3
langchain-chroma 
faiss-cpu 
numpy
langsmith==0.3.43
tenacity==9.1.2

#file reading
python-docx
PyMuPDF
pandas 


